`getAvailableFunds` fix
