+++ src/contract-logic/queries/getAuctions.ts
@@ -1,62 +1,19 @@
 import { Connection, PublicKey } from "@solana/web3.js"
 import { deserializeUnchecked } from "borsh"
+import { getAvailableFunds } from "./getAvailableFunds"
+import {
+  Bid,
+  NftData,
+  TokenData,
+  AuctionBaseConfig,
+  AuctionConfig,
+  AuctionBase,
+  Auction,
+} from "./types"
 import { LAMPORTS } from "../consts"
 import { AuctionPool, AuctionRootState, FrontendAuction, SCHEMA } from "../schema"
 import { padTo32Bytes } from "../utils/padTo32Bytes"
 import { parseAuctionId } from "../utils/parseAuctionId"
-
-export type Bid = {
-  bidderPubkey: PublicKey
-  amount: number
-}
-
-export type NFTData = {
-  type: "NFT"
-  name: string
-  symbol: string
-  uri: string
-  isRepeated: boolean
-}
-
-export type TokenData = {
-  type: "TOKEN"
-  decimals: number
-  mintAddress: PublicKey
-  perCycleAmount: number
-}
-
-export type AuctionBaseConfig = {
-  id: string
-  name: string
-  goalTreasuryAmount?: number
-  ownerPubkey: PublicKey
-}
-
-// TODO: customizable encore period? (The amount of time the highest bidder must be uncontested to win the cycle)
-export type AuctionConfig = AuctionBaseConfig & {
-  description: string
-  socials: string[]
-  asset: NFTData | TokenData
-  cyclePeriod: number
-  numberOfCycles: number
-  minBid: number
-  startTimestamp?: number
-}
-
-export type AuctionBase = AuctionBaseConfig & {
-  currentTreasuryAmount: number
-}
-
-export type Auction = AuctionConfig &
-  AuctionBase & {
-    availableTreasuryAmount,
-    bids: Bid[]
-    thisCycle: number
-    currentCycle: number
-    endTimestamp: number
-    isActive: boolean
-    isFrozen: boolean
-  }
 
 async function getAuctionPool(connection: Connection): Promise<AuctionPool> {
   const { getAuctionPoolPubkeyWasm } = await import("../../../wasm-factory")
@@ -103,8 +60,8 @@ export async function getAuctions(connection: Connection): Promise<Array<Auction
   return auctionBaseArray
 }
 
-export async function getAuction(id: string, n?: number): Promise<Auction> {
-  const { getAuctionWasm, getTreasuryWasm } = await import("../../../wasm-factory")
+export async function getAuction(auction_id: string, n?: number): Promise<Auction> {
+  const { getAuctionWasm } = await import("../../../wasm-factory")
   let cycle;
   if (n) {
     cycle = BigInt(n);
@@ -115,16 +72,14 @@ export async function getAuction(id: string, n?: number): Promise<Auction> {
   let auction;
   let availableLamports;
   try {
-    const auctionData: Uint8Array = await getAuctionWasm(id, cycle);
+    const auctionData: Uint8Array = await getAuctionWasm(auction_id, cycle);
     auction = deserializeUnchecked(
       SCHEMA,
       FrontendAuction,
       Buffer.from(auctionData),
     );
-
-    availableLamports = await getTreasuryWasm(id);
   } catch(error) {
-    console.log("wasm error:", error)
+    console.log("wasm error: ", error)
   }
 
   let asset: TokenData | NFTData
@@ -155,19 +110,28 @@ export async function getAuction(id: string, n?: number): Promise<Auction> {
     thisCycle = currentCycle;
   }
 
+  let bids = auction.cycleState.bidHistory
+      .map((bid) => ({ bidderPubkey: bid.bidderPubkey, amount: bid.bidAmount.toNumber() / LAMPORTS }))
+      .reverse();
+
+  let availableTreasuryAmount;
+  try {
+    availableTreasuryAmount = await getAvailableFunds(auction_id, bids[0])
+  } catch(error) {
+    console.log("wasm error: ", error)
+  }
+
   return {
-    id: id,
+    id: auction_id,
     name: parseAuctionId(Uint8Array.from(auction.rootState.auctionName)),
     description: auction.rootState.description.description,
     socials: auction.rootState.description.socials,
     goalTreasuryAmount,
-    availableTreasuryAmount: Number(availableLamports) / LAMPORTS,
+    availableTreasuryAmount,
     currentTreasuryAmount: auction.rootState.currentTreasury.toNumber() / LAMPORTS,
     ownerPubkey: auction.rootState.auctionOwner,
     asset: asset,
-    bids: auction.cycleState.bidHistory
-      .map((bid) => ({ bidderPubkey: bid.bidderPubkey, amount: bid.bidAmount.toNumber() / LAMPORTS }))
-      .reverse(),
+    bids, 
     cyclePeriod: auction.rootState.auctionConfig.cyclePeriod.toNumber(),
     thisCycle,
     currentCycle,
